10272: Renaming snapshot files to have `.js.snap` extension on CI
This operation is needed because on CI server we don't use Jest built-in
handling of TypeScript compilation. We compile tests to JS and then run
them. This causes that test are looking for the matching `.js.snap` file.

We can't use Jest built-in TypeScript support because of bugged coverage
reports generated in such configuration.